back/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
pydantic>=2.8.0,<3.0.0
langchain>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
faiss-cpu>=1.8.0,<2.0.0
numpy>=1.26.0,<3.0.0
rank-bm25>=0.2.2,<0.3.0
SQLAlchemy>=2.0.0,<3.0.0
PyMySQL>=1.1.0,<2.0.0
